# openapi.yaml
...
openapi: 3.1.0
info:
  title: openagentix API
  version: 0.1.0
  description: "Control node API of openagentix: agents, events, runs, approvals, audit, costs and RBAC."
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: Session or API token (`oax_...`)
    runToken:
      type: http
      scheme: bearer
      description: Signed run token (`oaxrt....`) of a worker node
    webhookSignature:
      type: apiKey
      in: header
      name: x-oax-signature
      description: HMAC-SHA256 signature
  schemas:
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: stable error code
        message:
          type: string
        details: {}
paths:
...
  /v1/event-sources:
    get:
      summary: List event sources
      tags:
        - events
      security:
        - bearer: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        name:
                          type: string
                        kind:
                          type: string
                          enum:
                            - webhook
                            - mail
                            - kafka
                            - cron
                        scheme:
                          type: string
                          enum:
                            - oax-v1
                            - github
                        secretRefs:
                          type: array
                          items:
                            type: string
                        agentId:
                          anyOf:
                            - type: string
                              format: uuid
                              pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                            - type: "null"
                        config:
                          type: object
                          propertyNames:
                            type: string
                          additionalProperties: {}
                        enabled:
                          type: boolean
                        createdAt:
                          type: string
                          description: ISO 8601 timestamp
                        ingestUrl:
                          type:
                            - string
                            - "null"
                      required:
                        - id
                        - name
                        - kind
                        - scheme
                        - secretRefs
                        - agentId
                        - config
                        - enabled
                        - createdAt
                        - ingestUrl
                      additionalProperties: false
                required:
                  - items
                additionalProperties: false
        "401":
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create an event source (secrets by reference)
      tags:
        - events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  pattern: ^[a-z][a-z0-9-]{0,62}$
                kind:
                  type: string
                  enum:
                    - webhook
                    - mail
                    - kafka
                    - cron
                scheme:
                  type: string
                  enum:
                    - oax-v1
                    - github
                secretRefs:
                  maxItems: 4
                  type: array
                  items:
                    type: string
                    minLength: 1
                agentId:
                  anyOf:
                    - type: string
                      format: uuid
                      pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                    - type: "null"
                config:
                  type: object
                  propertyNames:
                    type: string
                  additionalProperties: {}
                enabled:
                  type: boolean
              required:
                - name
                - kind
      security:
        - bearer: []
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  name:
                    type: string
                  kind:
                    type: string
                    enum:
                      - webhook
                      - mail
                      - kafka
                      - cron
                  scheme:
                    type: string
                    enum:
                      - oax-v1
                      - github
                  secretRefs:
                    type: array
                    items:
                      type: string
                  agentId:
                    anyOf:
                      - type: string
                        format: uuid
                        pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                      - type: "null"
                  config:
                    type: object
                    propertyNames:
                      type: string
                    additionalProperties: {}
                  enabled:
                    type: boolean
                  createdAt:
                    type: string
                    description: ISO 8601 timestamp
                  ingestUrl:
                    type:
                      - string
                      - "null"
                required:
                  - id
                  - name
                  - kind
                  - scheme
                  - secretRefs
                  - agentId
                  - config
                  - enabled
                  - createdAt
                  - ingestUrl
                additionalProperties: false
        "401":
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
                  details: {}
                required:
                  - error
                  - message
                additionalProperties: false
  /v1/event-sources/{id}:
    get:
      summary: Get an event source
      tags:
        - events
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: id
          required: true
      security:
        - bearer: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  name:
                    type: string
                  kind:
                    type: string
                    enum:
                      - webhook
                      - mail
                      - kafka
                      - cron
                  scheme:
                    type: string
                    enum:
                      - oax-v1
                      - github
                  secretRefs:
                    type: array
                    items:
                      type: string
                  agentId:
                    anyOf:
                      - type: string
                        format: uuid
                        pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                      - type: "null"
                  config:
                    type: object
                    propertyNames:
                      type: string
                    additionalProperties: {}
                  enabled:
                    type: boolean
                  createdAt:
                    type: string
                    description: ISO 8601 timestamp
                  ingestUrl:
                    type:
                      - string
                      - "null"
                required:
                  - id
                  - name
                  - kind
                  - scheme
                  - secretRefs
                  - agentId
                  - config
                  - enabled
                  - createdAt
                  - ingestUrl
                additionalProperties: false
        "401":
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      summary: Update an event source
      tags:
        - events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scheme:
                  type: string
                  enum:
                    - oax-v1
                    - github
                secretRefs:
                  maxItems: 4
                  type: array
                  items:
                    type: string
                    minLength: 1
                agentId:
                  anyOf:
                    - type: string
                      format: uuid
                      pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                    - type: "null"
                config:
                  type: object
                  propertyNames:
                    type: string
                  additionalProperties: {}
                enabled:
                  type: boolean
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: id
          required: true
      security:
        - bearer: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  name:
                    type: string
                  kind:
                    type: string
                    enum:
                      - webhook
                      - mail
                      - kafka
                      - cron
                  scheme:
                    type: string
                    enum:
                      - oax-v1
                      - github
                  secretRefs:
                    type: array
                    items:
                      type: string
                  agentId:
                    anyOf:
                      - type: string
                        format: uuid
                        pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                      - type: "null"
                  config:
                    type: object
                    propertyNames:
                      type: string
                    additionalProperties: {}
                  enabled:
                    type: boolean
                  createdAt:
                    type: string
                    description: ISO 8601 timestamp
                  ingestUrl:
                    type:
                      - string
                      - "null"
                required:
                  - id
                  - name
                  - kind
                  - scheme
                  - secretRefs
                  - agentId
                  - config
                  - enabled
                  - createdAt
                  - ingestUrl
                additionalProperties: false
        "401":
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete an event source (its events are kept)
      tags:
        - events
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: id
          required: true
      security:
        - bearer: []
      responses:
        "204":
          description: Default Response
        "401":
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/events:
    get:
      summary: List received events (newest first)
      tags:
        - events
      parameters:
        - schema:
            default: 50
            type: integer
            minimum: 1
            maximum: 200
          in: query
          name: limit
          required: false
        - schema:
            type: string
            maxLength: 200
          in: query
          name: cursor
          required: false
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: query
          name: sourceId
          required: false
      security:
        - bearer: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        sourceId:
                          anyOf:
                            - type: string
                              format: uuid
                              pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                            - type: "null"
                        cloudEventId:
                          type: string
                        type:
                          type: string
                        subject:
                          type:
                            - string
                            - "null"
                        receivedAt:
                          type: string
                          description: ISO 8601 timestamp
                        payload: {}
                      required:
                        - id
                        - sourceId
                        - cloudEventId
                        - type
                        - subject
                        - receivedAt
                        - payload
                      additionalProperties: false
                  nextCursor:
                    type:
                      - string
                      - "null"
                required:
                  - items
                  - nextCursor
                additionalProperties: false
        "401":
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/events/{id}:
    get:
      summary: Get an event
      tags:
        - events
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: id
          required: true
      security:
        - bearer: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  sourceId:
                    anyOf:
                      - type: string
                        format: uuid
                        pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                      - type: "null"
                  cloudEventId:
                    type: string
                  type:
                    type: string
                  subject:
                    type:
                      - string
                      - "null"
                  receivedAt:
                    type: string
                    description: ISO 8601 timestamp
                  payload: {}
                required:
                  - id
                  - sourceId
                  - cloudEventId
                  - type
                  - subject
                  - receivedAt
                  - payload
                additionalProperties: false
        "401":
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
servers:
  - url: http://localhost:8080
    description: local development